feat: added filtering to scripts
