+++ api/v1/testkube.yaml
@@ -33,6 +33,9 @@ paths:
       summary: "Create new test"
       description: "Create new test action"
       operationId: createTest
+      parameters:
+        - $ref: "#/components/parameters/Tags"
+        - $ref: "#/components/parameters/TextSearch"
       requestBody:
         $ref: "#/components/schemas/TestUpsertRequest"
       responses:
@@ -286,6 +289,7 @@ paths:
       operationId: listScripts
       parameters:
         - $ref: "#/components/parameters/Tags"
+        - $ref: "#/components/parameters/TextSearch"
       responses:
         200:
           description: "successful operation"
